requirements.txt
Django==3.2.5
openpyxl==3.0.7
djangorestframework==3.12.4
celery==5.1.2
redis==3.5.3